+++ netbox_catalyst_center/__init__.py
@@ -10,7 +10,7 @@ import logging
 from django.db.models.signals import post_migrate
 from netbox.plugins import PluginConfig
 
-__version__ = "1.3.4"
+__version__ = "1.3.5"
 
 logger = logging.getLogger(__name__)